Arduino code for controlling the aircraft.
This software is for use with the Arduino Mega 2560 and the Adafruit
10DOF Breakout. The intended use of this software is the control and
stabilization of a Y-3 configuration tricopter. Static variable,
manually tuned alpha-beta kalman filters are used to fuse the gyroscope
and accelerometer raw data to determine craft orientation. A moving
average filter is used on each to remove high frequency vibrations
caused by the engines (a mechanical vibration isolator is necessary to
fully stabilize the readout).